+++ frontend/package.json
@@ -13,7 +13,8 @@
   "dependencies": {
     "firebase": "^12.1.0",
     "react": "^19.1.1",
-    "react-dom": "^19.1.1"
+    "react-dom": "^19.1.1",
+    "react-router-dom": "^7.8.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.33.0",
