+++ frontend/src/App.css
@@ -40,3 +40,17 @@
 .read-the-docs {
   color: #888;
 }
+
+.home-page {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 80px;
+}
+
+.feature-buttons button {
+  margin: 10px;
+  padding: 15px 30px;
+  font-size: 18px;
+  cursor: pointer;
+}
