+++ frontend/src/Chatbot.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Chatbot: React.FC = () => {
+    return (
+        <div>
+            <h2>Chatbot</h2>
+        </div>
+    );
+};
+
+export default Chatbot;
